+++ src/control_plane/shared_infrastructure/gotenberg/reconcile_prepared_gotenberg_instance.rs
@@ -39,5 +39,6 @@ where
         shared.container().metadata(),
         None,
         logical,
+        shared.health(),
     ))
 }