+++ src/control_plane/shared_infrastructure/mailpit/reconcile_prepared_mailpit_instance.rs
@@ -39,5 +39,6 @@ where
             .volume()
             .map(|volume| (volume.volume().name(), volume.volume().metadata())),
         logical,
+        shared.health(),
     ))
 }