+++ src/control_plane/shared_infrastructure/mongodb/reconcile_prepared_mongodb_instance.rs
@@ -64,5 +64,6 @@ where
             .volume()
             .map(|volume| (volume.volume().name(), volume.volume().metadata())),
         logical,
+        shared.health(),
     ))
 }