+++ src/control_plane/shared_infrastructure/mysql/reconcile_prepared_mysql_instance.rs
@@ -60,5 +60,6 @@ where
             .volume()
             .map(|volume| (volume.volume().name(), volume.volume().metadata())),
         logical,
+        shared.health(),
     ))
 }