+++ src/control_plane/shared_infrastructure/rabbitmq/reconcile_prepared_rabbitmq_instance.rs
@@ -44,5 +44,6 @@ where
             .volume()
             .map(|volume| (volume.volume().name(), volume.volume().metadata())),
         logical,
+        shared.health(),
     ))
 }